fix(rust): Wave 5-E 列名漂移全仓修复 + clippy 清零
## 列名漂移(9 处)

1. rpg-routes/src/timeline.rs
   - `platform_saves` → `game_saves`(表不存在)
   - `script_timeline_anchors.phase/label/anchor_turn`
     → `story_phase/story_time_label/chapter_min/chapter_max`
   - `save_phase_digests.label/turn`
     → `phase_label/turn_start`

2. rpg-agents/src/worldbook_agent.rs
   - `script_timeline_anchors.phase/time_label`
     → `story_phase/story_time_label`

3. rpg-retrieval/src/lib.rs
   - `chapter_facts.events_json`(不存在)
     → `events::text AS events_json`

4. rpg-llm/src/registry.rs
   - `model_apis.id` → `api_id`(PK 列名漂移)
   - `model_models` 表不存在 → `model_entries`
   - `model_entries.id(text)/capabilities(text[])` 对齐实际 schema
   - 价格列不在 DB schema,定价设 None

5. rpg-db/src/repos/worldbook_entries.rs
   - 全量重写:旧版用 `key/aliases/comment/save_id/user_id/tags`
     均不存在于 migration 001 的 worldbook_entries 表
   - 新版对齐 `title/keys/regex_keys/...` 真实列
   - `list_for_save` 表无 save_id 列,改为返回空列表(保留签名兼容)

6. rpg-context/src/providers/script_phase_anticipation.rs
   - `chapter_min/chapter_max` 类型 i64 → i32(对应 integer 列)

## 新增 migration 023

- `023_phase_digests.sql`: 创建 script 级 `phase_digests` 表
  (对应 Python aggregate_phase_digests.py 写入的表,此前 Rust 代码
   引用该表但 DB 里不存在,运行时 relation "phase_digests" does not exist)
- migrations.rs 注册 SQL_023 + MIGRATIONS[id=23]

## clippy 清零(6 warning → 0)

- rpg-state/src/rules_outcome.rs: 去掉 4 处无效 `.map_err(Into::into)`
  (set_path/append_path 已返回 Result<_, StateError>)
- rpg-agents/src/worldbook_agent.rs: 2 处冗余闭包
  `.map(|x| Value::from(x))` → `.map(Value::from)`

cargo check --workspace: 0 error
cargo clippy --workspace --all-targets -- -D warnings: 0 warning
cargo test --workspace: passed 219 / failed 6(均为预存 upload 测试,不回退)

+++ rust/crates/rpg-context/src/providers/script_phase_anticipation.rs
@@ -117,7 +117,8 @@ async fn load_script_lookahead(
         None => return Ok(Vec::new()),
     };
 
-    let mut chapter_threshold: i64 = 0;
+    // chapter_min/chapter_max 在 phase_digests 表中是 integer(i32)。
+    let mut chapter_threshold: i32 = 0;
     if !current_phase.is_empty() {
         let row = sqlx::query(
             "select chapter_max from phase_digests \
@@ -128,7 +129,7 @@ async fn load_script_lookahead(
         .fetch_optional(pool)
         .await?;
         if let Some(row) = row {
-            chapter_threshold = row.try_get::<i64, _>("chapter_max").unwrap_or(0);
+            chapter_threshold = row.try_get::<i32, _>("chapter_max").unwrap_or(0);
         }
     }
 
@@ -151,8 +152,8 @@ async fn load_script_lookahead(
         .into_iter()
         .map(|row| {
             let phase_label: Option<String> = row.try_get("phase_label").ok();
-            let chapter_min: i64 = row.try_get("chapter_min").unwrap_or(0);
-            let chapter_max: i64 = row.try_get("chapter_max").unwrap_or(0);
+            let chapter_min: i32 = row.try_get("chapter_min").unwrap_or(0);
+            let chapter_max: i32 = row.try_get("chapter_max").unwrap_or(0);
             let summary: Option<String> = row.try_get("summary").ok();
             let key_events: Value = row.try_get("key_events").unwrap_or(Value::Null);
             let key_locations: Value = row.try_get("key_locations").unwrap_or(Value::Null);
